hooked up edit student form from result, however, a new bug arises and need further investigation
the course and rotation is funky! cna value doesnt look like its
behaving right.

only course displayed is CNA doesnt matter what I choose, but it is
reflected in the rotation abbreviations,

but i kind of have a clue, its the reocordForEdit? recordForEdit.course:
course,

its always going to be recordForEdit, and the acutal course state is
ignoreed