changed all inputs from input2 to input, and deleted the old implementation.
Next, I have decided to keep queryform as is. Since its fields do not
exceed 3 per line, and 17 components max, I have decided that since this
is an optimization edit. And queryForm performance hit is not as bad as
studentForm, since studentForm has so many inputs and components
per-render.

In this case, simply changing the base component of searchBar to the
base MuiTextField, which voids the useInputValue custom hook, I am able
to keep the current implementation. (which would proof to be more
complicated to implement, and would feel more complicated overall) the
only reason why the new pattern was implemented was so it can help boost
performance in large forms such as studentforms. As mentioned
previously, the queryForm is really not a big component itself.

Now, next I will further boost performance, by separating states of
query and queryform (maybe even another container component that will
call the custom hook for the form. More than likely will take out the
useReducers inside useQueryForm, and just use useState instead.