implemented logout,
changed some of the test, to mock axios api in login, need to also
remember to test logout functionailities as well.

also, added some notes for the logic inside persistLogin espcially when
it comes to what goes on inside a refresh call.

Also, added a refresh call try catch async block inside useSigninforms() to check everytime /
login page is hit, is there a refresh token, if it is (inside
useRefresToekn hook, it also sets the user and authed, and will then
redirect to /query. But if it was no good inside refresh, then caller
needs to handle error and unset authed and user, will also stay at the /
login page.

need to think about finding the right pattern to do this. where should
we setUser and setAuthed? and where should we handle these errors, might
need to change useRefreshToken hook to only call refresh api, but not
handle internal states, that should be left inside each caller, imo.